import components from src in pages
