+++ pages/team.js
@@ -1,242 +1,15 @@
 import React from "react";
+import { NextSeo } from "next-seo";
+import Committee from "../src/client/components/committee";
 
 export default function TeamPage() {
   return (
-    <div className="content-block">
-      <div className="text-l text-b">Committee</div>
-      <div className="text-s margin-top-s">
-        Our shared love of art, music, and design
-      </div>
-      <div className="page-container margin-top">
-        <div className="text-m">Advisors</div>
-        <a
-          target="_blank"
-          className="text-s text-grey pointer"
-          href="https://y.at/%F0%9F%91%89%F0%9F%8E%B1%F0%9F%95%B3"
-        >
-          Tim Kang (illestrater)
-        </a>
-        <div className="text-s">Sevens Developer & paradigm pusher</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://twitter.com/darlington"
-        >
-          Mike Darlington
-        </a>
-        <div className="text-s">
-          CEO of Monstercat music label & captain NFT gremlin
-        </div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.slimesunday.com"
-        >
-          Devin Dube
-        </a>
-        <div className="text-s">
-          Digital Marketing @ Slime Sunday & art lover
-        </div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://twitter.com/EthArtX"
-        >
-          Larry Saggese
-        </a>
-        <div className="text-s">NFT podcaster & communications guru</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://3lau.com/"
-        >
-          Justin Blau (3LAU)
-        </a>
-        <div className="text-s">Music producer & NFT innovator</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.instagram.com/yesladypheonix/"
-        >
-          Lady PheOnix
-        </a>
-        <div className="text-s">
-          Co-Founder of Universe Contemporary & leading crypto art evangelist
-        </div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://twitter.com/axieking"
-        >
-          OhhShiny
-        </a>
-        <div className="text-s">Traditional fine art & NFT collector</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://twitter.com/pablorfraile"
-        >
-          Pablo
-        </a>
-        <div className="text-s">
-          Co-Founder of Museum of Crypto Art & NFT collector
-        </div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://twitter.com/LordTylerWard"
-        >
-          Tyler Ward
-        </a>
-        <div className="text-s">Founder of Barn Bridge & DeFi degen</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://y.at/%E2%AD%90%F0%9F%8E%AC%E2%AD%90"
-        >
-          Brandon Kang
-        </a>
-        <div className="text-s">Professional videographer & NFT collector</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="http://justinmaller.com/"
-        >
-          Justin Maller
-        </a>
-        <div className="text-s">Digital Artist & CCO of Deviantart</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://parishilton.com/"
-        >
-          Paris Hilton
-        </a>
-        <div className="text-s">Entrepreneur & NFT and social pioneer</div>
-        <a
-          target="_blank"
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.beeple-crap.com/"
-        >
-          Mike Winkelmann (beeple)
-        </a>
-        <div className="text-s">
-          Renowned digital art director & professional hair trimmer
-        </div>
-        <a target="_blank" className="text-s margin-top-s text-grey pointer">
-          Keira
-        </a>
-        <div className="text-s">Software engineer & database rancher</div>
-        <div className="text-m margin-top">Artists</div>
-        <div
-          className="text-s margin-top-s text-grey pointer"
-          href="https://nessgraphics.com/"
-        >
-          NessGraphics
-        </div>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.instagram.com/fvckrender"
-          target="_blank"
-        >
-          FVCKRENDER
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://aeforiadesign.com/"
-          target="_blank"
-        >
-          aeforia
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://thankyoux.com/"
-          target="_blank"
-        >
-          ThankYouX
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://maalavidaa.com/"
-          target="_blank"
-        >
-          Maalavidaa
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="http://www.blakekathryn.com/"
-          target="_blank"
-        >
-          Blake Kathryn
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://slimesunday.com/"
-          target="_blank"
-        >
-          slimesunday
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.instagram.com/toomuchlag"
-          target="_blank"
-        >
-          toomuchlag
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.victormosquera.com/"
-          target="_blank"
-        >
-          Victor Mosquera
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://smeccea.com/"
-          target="_blank"
-        >
-          smeccea
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.billelis.com/"
-          target="_blank"
-        >
-          Billelis
-        </a>
-        <br />
-        <a
-          className="text-s margin-top-s text-grey pointer"
-          href="https://www.behance.net/Filiphds"
-          target="_blank"
-        >
-          Filip Hodas
-        </a>
-        <br />
-        <br />
-        <div className="text-m text-b">Partners</div>
-        <div className="text-s margin-top">
-          <a
-            className="text-s text-grey pointer"
-            href="https://www.numomo.com/"
-            target="_blank"
-          >
-            Numomo Creative Agency
-          </a>
-          <div className="text-s">World's First Creative NFT Agency</div>
-        </div>
-        <br />
-      </div>
-    </div>
+    <>
+      <NextSeo
+        title="Seven's Foundation"
+        description="Our shared love of art, music, and design. Sevens Foundation was created by Tim Kang through the heart of Mike Darlington. We are a global and 501(c)3 non-profit organization with a mission to spotlight emerging artists and share their creativity with the world."
+      />
+      <Committee />
+    </>
   );
 }
