+++ pages/testimony.js
@@ -1,6 +1,15 @@
 import React from "react";
+import { NextSeo } from "next-seo";
 import Testimony from "../src/client/components/testimony";
 
 export default function TestimonyPage() {
-  return <Testimony />;
+  return (
+    <>
+      <NextSeo
+        title="Testimonials"
+        description="Artist lives are changing drastically through NFTs."
+      />
+      <Testimony />
+    </>
+  );
 }
